feat: wip tx cli
